modify: add namespace to model clases
