+++ public/index.php
@@ -20,8 +20,6 @@ if (is_readable($dot_env)) {
 
 Resque::setBackend('localhost:6379');
 
-require __DIR__ . '/../src/models.php';
-
 ini_set('session.gc_maxlifetime', 86400);
 session_start();
