+++ src/Models/Genre.php
@@ -2,11 +2,13 @@
 /**
  * Created by PhpStorm.
  * User: yagihash
- * Date: 2017/09/30
- * Time: 18:41
+ * Date: 2017/09/23
+ * Time: 16:46
  */
 
-class Mode
+namespace SenbonXSS\Models;
+
+class Genre
 {
     /**
      * @var int
@@ -16,22 +18,16 @@ class Mode
      * @var string
      */
     protected $name = '';
-    /**
-     * @var string
-     */
-    protected $handler = '';
 
     /**
-     * Mode constructor.
+     * Genre constructor.
      * @param int $id
      * @param string $name
-     * @param string $handler
      */
-    function __construct(int $id, string $name, string $handler)
+    function __construct(int $id, string $name)
     {
         $this->id = $id;
-        $this->name =$name;
-        $this->handler = $handler;
+        $this->name = $name;
     }
 
     /**
@@ -48,13 +44,5 @@ class Mode
     function getName(): string
     {
         return $this->name;
-    }
-
-    /**
-     * @return string
-     */
-    function getHandler(): string
-    {
-        return $this->handler;
     }
 }
