+++ src/Models/Modes.php
@@ -2,24 +2,28 @@
 /**
  * Created by PhpStorm.
  * User: yagihash
- * Date: 2017/09/26
- * Time: 15:11
+ * Date: 2017/09/30
+ * Time: 18:41
  */
 
-class Genres
+namespace SenbonXSS\Models;
+
+use PDO;
+
+class Modes
 {
     /**
      * @param PDO $pdo
-     * @return Genre[]
+     * @return Mode[]
      */
-    function fetchAllGenres(PDO $pdo): array
+    function fetchAllModes(PDO $pdo): array
     {
         $ret = [];
-        $q = 'SELECT id, name FROM genres';
+        $q = 'SELECT id, name, handler FROM stagemode';
         $r = $pdo->query($q);
         $rows = $r->fetchAll();
         foreach ($rows as $row) {
-            $ret[] = new Genre($row['id'], $row['name']);
+            $ret[] = new Mode($row['id'], $row['name'], $row['handler']);
         }
         return $ret;
     }
