--- src/models.php
@@ -1,6 +1,0 @@
-<?php
-$models = glob(__DIR__ . '/models/*.php');
-
-foreach($models as $path) {
-    require_once($path);
-}
